requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
lxml>=4.9.0

